Simplify cloud examples into a single project
