--- examples/cloud-grpc-client-consul/src/main/resources/application.yml
@@ -1,12 +1,0 @@
-server:
-  port: 8080
-spring:
-  application:
-    name: cloud-grpc-client
-grpc:
-  client:
-    cloud-grpc-server:
-      address: 'discovery:///cloud-grpc-server'
-      enableKeepAlive: true
-      keepAliveWithoutCalls: true
-      negotiationType: plaintext
