--- examples/cloud-grpc-client-zookeeper/src/main/java/net/devh/boot/grpc/examples/cloud/client/GrpcClientService.java
@@ -1,50 +1,0 @@
-/*
- * Copyright (c) 2016-2020 Michael Zhang <yidongnan@gmail.com>
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
- * documentation files (the "Software"), to deal in the Software without restriction, including without limitation the
- * rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to
- * permit persons to whom the Software is furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all copies or substantial portions of the
- * Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE
- * WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
- * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
- * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-
-package net.devh.boot.grpc.examples.cloud.client;
-
-import org.springframework.stereotype.Service;
-
-import io.grpc.StatusRuntimeException;
-import lombok.extern.slf4j.Slf4j;
-import net.devh.boot.grpc.client.inject.GrpcClient;
-import net.devh.boot.grpc.examples.lib.HelloReply;
-import net.devh.boot.grpc.examples.lib.HelloRequest;
-import net.devh.boot.grpc.examples.lib.SimpleGrpc.SimpleBlockingStub;
-
-/**
- * @author zhaochunlin (946599275@qq.com)
- * @since 2021/1/9
- */
-@Service
-@Slf4j
-public class GrpcClientService {
-
-    @GrpcClient("cloud-grpc-server-zookeeper")
-    private SimpleBlockingStub simpleStub;
-
-    public String sendMessage(final String name) {
-        try {
-            final HelloReply response = this.simpleStub.sayHello(HelloRequest.newBuilder().setName(name).build());
-            return response.getMessage();
-        } catch (final StatusRuntimeException e) {
-            log.error(e.getMessage());
-            return "FAILED with " + e.getStatus().getCode();
-        }
-    }
-
-}
