+++ examples/cloud-grpc-client/src/main/java/net/devh/boot/grpc/examples/cloud/client/CloudClientApplication.java
