--- examples/cloud-grpc-client/src/main/java/net/devh/boot/grpc/examples/cloud/client/CloudGrpcClientApplication.java
@@ -1,38 +1,0 @@
-/*
- * Copyright (c) 2016-2021 Michael Zhang <yidongnan@gmail.com>
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
- * documentation files (the "Software"), to deal in the Software without restriction, including without limitation the
- * rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to
- * permit persons to whom the Software is furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all copies or substantial portions of the
- * Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE
- * WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
- * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
- * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-
-package net.devh.boot.grpc.examples.cloud.client;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
-
-/**
- * @author Michael (yidongnan@gmail.com)
- * @since 2016/11/8
- */
-@EnableEurekaClient
-@EnableDiscoveryClient
-@SpringBootApplication
-public class CloudGrpcClientApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(CloudGrpcClientApplication.class, args);
-    }
-
-}
