+++ examples/cloud-grpc-client/src/main/resources/application-consul.yml
