+++ examples/cloud-grpc-client/src/main/resources/application.yml
@@ -3,23 +3,10 @@ server:
 spring:
   application:
     name: cloud-grpc-client
-  sleuth:
-    sampler:
-      probability: 1
-eureka:
-  instance:
-    prefer-ip-address: true
-    status-page-url-path: /actuator/info
-    health-check-url-path: /actuator/health
-    instanceId: ${spring.application.name}:${vcap.application.instance_id:${spring.application.instance_id:${random.value}}}
-  client:
-    register-with-eureka: true
-    fetch-registry: true
-    service-url:
-      defaultZone: http://localhost:8761/eureka/
 grpc:
   client:
     cloud-grpc-server:
+      address: 'discovery:///cloud-grpc-server'
       enableKeepAlive: true
       keepAliveWithoutCalls: true
       negotiationType: plaintext
