--- examples/cloud-grpc-server-consul/src/main/resources/application.yml
@@ -1,10 +1,0 @@
-server:
-  port: 58080
-
-spring:
-  application:
-    name: cloud-grpc-server
-
-grpc:
-  server:
-    port: 58081
