--- examples/cloud-grpc-server-zookeeper/build.gradle
@@ -1,16 +1,0 @@
-plugins {
-    id 'org.springframework.boot'
-}
-
-dependencies {
-    implementation 'org.springframework.cloud:spring-cloud-starter-zookeeper-discovery'
-    implementation 'org.springframework.boot:spring-boot-starter-actuator'
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-    implementation project(':examples:grpc-lib')
-    implementation project(':grpc-server-spring-boot-starter') // replace to implementation "net.devh:grpc-server-spring-boot-starter:${springBootGrpcVersion}"
-
-    testImplementation('org.springframework.boot:spring-boot-starter-test'){
-        exclude module: 'junit-vintage-engine'
-        exclude module: 'junit'
-    }
-}
