--- examples/cloud-grpc-server-zookeeper/src/main/resources/application.yml
@@ -1,15 +1,0 @@
-server:
-  port: 8081
-
-spring:
-  application:
-    name: cloud-grpc-server-zookeeper
-  cloud:
-    zookeeper:
-      discovery:
-        register: true
-      connect-string: localhost:2181
-
-grpc:
-  server:
-    port: 9999
