+++ examples/cloud-grpc-server/src/main/resources/application-consul.yml
