+++ examples/cloud-grpc-server/src/main/resources/application-zookeeper.yml
@@ -1,0 +1,6 @@
+#spring:
+#  cloud:
+#    zookeeper:
+#      connect-string: localhost:2181
+#      discovery:
+#        register: true
