+++ examples/cloud-grpc-server/src/main/resources/application.yml
@@ -1,20 +1,10 @@
+server:
+  port: 58080
+
 spring:
   application:
     name: cloud-grpc-server
-  sleuth:
-    sampler:
-      probability: 1
-server:
-  port: 0
+
 grpc:
   server:
     port: 0
-eureka:
-  instance:
-    prefer-ip-address: true
-    instanceId: ${spring.application.name}:${vcap.application.instance_id:${spring.application.instance_id:${random.value}}}
-  client:
-    register-with-eureka: true
-    fetch-registry: true
-    service-url:
-      defaultZone: http://localhost:8761/eureka/
