+++ testExamples.sh
@@ -1,9 +1,13 @@
 #!/bin/bash
 set -e # Fail on error
 trap "trap - SIGTERM && kill -- -$$" SIGINT SIGTERM EXIT # Kill subprocesses on exit
-./gradlew clean --console=plain
-./gradlew build --console=plain
-sleep 2s
+
+build() {
+	echo "Building project"
+	./gradlew clean --console=plain
+	./gradlew build --console=plain
+	sleep 2s
+}
 
 ## Local
 localTest() {
@@ -46,30 +50,48 @@ localTest() {
 }
 
 ## Cloud-Eureka
-cloudEurekaTest() {
-	echo "Starting Cloud Eureka test"
+cloudTest() {
+	echo "Starting Cloud $1 test"
 
 	# Run environment
-	./gradlew :example:cloud-eureka-server:bootRun -x jar -x classes --console=plain &
-	EUREKA=$!
-	sleep 10s # Wait for the server to start
-
-	mkdir -p zipkin
-	cd zipkin
-	echo "*" > .gitignore
-	if [ ! -f zipkin.jar ]; then
-		curl -sSL https://zipkin.io/quickstart.sh | bash -s
+	if [[ "$1" = "consul" ]]; then
+		CONSUL=`docker run --name=consul -d --rm -p 8500:8500 consul`
+		stopCloudEnv() {
+			echo "Stopping consul server"
+			docker stop $CONSUL
+		}
+	elif [[ "$1" == "eureka" ]]; then
+		./gradlew :example:cloud-eureka-server:bootRun -x jar -x classes --console=plain &
+		EUREKA=$!
+		stopCloudEnv() {
+			echo "Stopping eureka server"
+			kill -s TERM $EUREKA
+		}
+	elif [[ "$1" = "nacos" ]]; then
+		NACOS=`docker run --env MODE=standalone --name nacos -d --rm -p 8848:8848 nacos/nacos-server`
+		stopCloudEnv() {
+			echo "Stopping nacos server"
+			docker stop $NACOS
+		}
 	fi
-	java -jar zipkin.jar &
-	ZIPKIN=$!
 	sleep 10s # Wait for the server to start
-	cd ..
 
-	./gradlew :example:cloud-grpc-server:bootRun -x jar -x classes --console=plain &
+#	mkdir -p zipkin
+#	cd zipkin
+#	echo "*" > .gitignore
+#	if [ ! -f zipkin.jar ]; then
+#		curl -sSL https://zipkin.io/quickstart.sh | bash -s
+#	fi
+#	java -jar zipkin.jar &
+#	ZIPKIN=$!
+#	sleep 10s # Wait for the server to start
+#	cd ..
+
+	./gradlew -Pdiscovery=$1 :example:cloud-grpc-server:bootRun -x jar -x classes --console=plain &
 	CLOUD_SERVER=$!
 	sleep 10s # Wait for the server to start
 
-	./gradlew :example:cloud-grpc-client:bootRun -x jar -x classes --console=plain &
+	./gradlew -Pdiscovery=$1 :example:cloud-grpc-client:bootRun -x jar -x classes --console=plain &
 	CLOUD_CLIENT=$!
 	sleep 30s # Wait for the client to start and the server to be ready
 	sleep 60s # Wait for the discovery service to refresh
@@ -89,7 +111,7 @@ cloudEurekaTest() {
 	sleep 1s # Wait for the shutdown logs to pass
 
 	# and restart server
-	./gradlew :example:cloud-grpc-server:bootRun -x jar -x classes --console=plain &
+	./gradlew -Pdiscovery=$1 :example:cloud-grpc-server:bootRun -x jar -x classes --console=plain &
 	CLOUD_SERVER=$!
 	sleep 30s # Wait for the server to start
 	sleep 60s # Wait for the discovery service to refresh
@@ -105,8 +127,8 @@ cloudEurekaTest() {
 
 	# Shutdown
 	echo "Triggering shutdown"
-	kill -s TERM $EUREKA
-	kill -s TERM $ZIPKIN
+	stopCloudEnv
+	# kill -s TERM $ZIPKIN
 	kill -s TERM $CLOUD_SERVER
 	kill -s TERM $CLOUD_CLIENT
 	sleep 1s # Wait for the shutdown logs to pass
@@ -114,11 +136,11 @@ cloudEurekaTest() {
 	# Verify part 1
 	if [ "$RESPONSE" = "$EXPECTED" ]; then
 		echo "#------------------------------------#"
-		echo "| Cloud Eureka example part 1 works! |"
+		echo "| Cloud $1 example part 1 works! |"
 		echo "#------------------------------------#"
 	else
 		echo "#-------------------------------------#"
-		echo "| Cloud Eureka example part 1 failed! |"
+		echo "| Cloud $1 example part 1 failed! |"
 		echo "#-------------------------------------#"
 		exit 1
 	fi
@@ -126,94 +148,14 @@ cloudEurekaTest() {
 	# Verify part 2
 	if [ "$RESPONSE2" = "$EXPECTED" ]; then
 		echo "#------------------------------------#"
-		echo "| Cloud Eureka example part 2 works! |"
+		echo "| Cloud $1 example part 2 works! |"
 		echo "#------------------------------------#"
 	else
 		echo "#-------------------------------------#"
-		echo "| Cloud Eureka example part 2 failed! |"
+		echo "| Cloud $1 example part 2 failed! |"
 		echo "#-------------------------------------#"
 		exit 1
 	fi
-}
-
-## Cloud-Nacos
-cloudNacosTest() {
-    echo "Starting Cloud Nacos test"
-
-    # Run environment
-    docker pull nacos/nacos-server
-    NACOS=`docker run --env MODE=standalone --name nacos -d -p 8848:8848 nacos/nacos-server`
-    sleep 10s # Wait for the nacos server to start
-
-    ./gradlew :example:cloud-grpc-server-nacos:bootRun -x jar -x classes --console=plain &
-    CLOUD_SERVER=$!
-    sleep 10s # Wait for the server to start
-
-    ./gradlew :example:cloud-grpc-client-nacos:bootRun -x jar -x classes --console=plain &
-    CLOUD_CLIENT=$!
-    sleep 30s # Wait for the client to start and the server to be ready
-    sleep 60s # Wait for the discovery service to refresh
-
-    # Test
-    RESPONSE=$(curl -s localhost:8080/)
-    echo "Response:"
-    echo "$RESPONSE"
-    EXPECTED=$(echo -e "Hello ==> Michael")
-    echo "Expected:"
-    echo "$EXPECTED"
-    sleep 1s # Give the user a chance to look at the result
-
-    # Crash server
-    kill -s TERM $CLOUD_SERVER
-    echo "The server crashed (expected)"
-    sleep 1s # Wait for the shutdown logs to pass
-
-    # and restart server
-    ./gradlew :example:cloud-grpc-server-nacos:bootRun -x jar -x classes --console=plain &
-    CLOUD_SERVER=$!
-    sleep 30s # Wait for the server to start
-    sleep 60s # Wait for the discovery service to refresh
-    
-    # Test again
-    RESPONSE2=$(curl -s localhost:8080/)
-    echo "Response:"
-    echo "$RESPONSE2"
-    EXPECTED=$(echo -e "Hello ==> Michael")
-    echo "Expected:"
-    echo "$EXPECTED"
-    sleep 1s # Give the user a chance to look at the result
-
-    # Shutdown
-    echo "Triggering shutdown"
-    docker stop $NACOS
-    docker rm -f $NACOS
-    kill -s TERM $CLOUD_SERVER
-    kill -s TERM $CLOUD_CLIENT
-    sleep 1s # Wait for the shutdown logs to pass
-
-    # Verify part 1
-    if [ "$RESPONSE" = "$EXPECTED" ]; then
-        echo "#-----------------------------------#"
-        echo "| Cloud Nacos example part 1 works! |"
-        echo "#-----------------------------------#"
-    else
-        echo "#------------------------------------#"
-        echo "| Cloud Nacos example part 1 failed! |"
-        echo "#------------------------------------#"
-        exit 1
-    fi
-
-    # Verify part 2
-    if [ "$RESPONSE2" = "$EXPECTED" ]; then
-        echo "#-----------------------------------#"
-        echo "| Cloud Nacos example part 2 works! |"
-        echo "#-----------------------------------#"
-    else
-        echo "#------------------------------------#"
-        echo "| Cloud Nacos example part 2 failed! |"
-        echo "#------------------------------------#"
-        exit 1
-    fi
 }
 
 ## Security Basic Auth
@@ -297,8 +239,10 @@ securityBearerAuthTest() {
 }
 
 ## Tests
+build
 localTest
-cloudEurekaTest
-cloudNacosTest
+#cloudTest consul
+cloudTest eureka
+cloudTest nacos
 securityBasicAuthTest
 #securityBearerAuthTest
